+++ src/array.mjs
@@ -41,3 +41,4 @@ export { default as array_walk_recursive } from './array/array_walk_recursive.mj
 export { default as in_array } from './array/in_array.mjs';
 export { default as key_exists } from './array/key_exists.mjs';
 export { default as range } from './array/range.mjs';
+export { default as shuffle } from './array/shuffle.mjs';